+++ example/lib/transitions.dart
@@ -11,90 +11,142 @@ class Info {
 
 List<Info> kTransitions = <Info>[
   Info(
-    'Bottom to Top Fade',
-    (_) => BottomToTopFadeTranzition(child: const Page2()),
-  ),
-  Info(
-    'Bottom to Top Join',
-    (BuildContext context) => BottomToTopJoinTranzition(
-      child: const Page2(),
-      currentChild: context.widget,
-    ),
-  ),
-  Info(
-    'Bottom to Top Pop',
-    (BuildContext context) => BottomToTopPopTranzition(
-      child: const Page2(),
-      currentChild: context.widget,
-    ),
-  ),
-  Info(
-    'Bottom to Top',
-    (_) => BottomToTopTranzition(
-      child: const Page2(),
-    ),
-  ),
-  Info(
     'Fade',
     (_) => FadeTranzition(
       child: const Page2(),
     ),
   ),
   Info(
-    'Left to Right Fade',
-    (_) => LeftToRightFadeTranzition(
+    'Fade (bottom to top)',
+    (_) => FadeTranzition.bottomToTop(
       child: const Page2(),
     ),
   ),
   Info(
-    'Left to Right Join',
-    (BuildContext context) => LeftToRightJoinTranzition(
+    'Fade (left to right)',
+    (_) => FadeTranzition.leftToRight(
+      child: const Page2(),
+    ),
+  ),
+  Info(
+    'Fade (right to left)',
+    (_) => FadeTranzition.rightToLeft(
+      child: const Page2(),
+    ),
+  ),
+  Info(
+    'Fade (top to bottom)',
+    (_) => FadeTranzition.topToBottom(
+      child: const Page2(),
+    ),
+  ),
+  Info(
+    'Join (bottom to top)',
+    (BuildContext context) => JoinTranzition.bottomToTop(
       child: const Page2(),
       currentChild: context.widget,
     ),
   ),
   Info(
-    'Left To Right Pop',
-    (BuildContext context) => LeftToRightPopTranzition(
+    'Join (left to right)',
+    (BuildContext context) => JoinTranzition.leftToRight(
       child: const Page2(),
       currentChild: context.widget,
     ),
   ),
   Info(
-    'Left to Right',
-    (_) => LeftToRightTranzition(
-      child: const Page2(),
-    ),
-  ),
-  Info(
-    'Right to Left Fade',
-    (_) => RightToLeftFadeTranzition(
-      child: const Page2(),
-    ),
-  ),
-  Info(
-    'Right to Left Join',
-    (BuildContext context) => RightToLeftJoinTranzition(
+    'Join (right to left)',
+    (BuildContext context) => JoinTranzition.rightToLeft(
       child: const Page2(),
       currentChild: context.widget,
     ),
   ),
   Info(
-    'Right to Left Pop',
-    (BuildContext context) => RightToLeftPopTranzition(
+    'Join (top to bottom)',
+    (BuildContext context) => JoinTranzition.topToBottom(
       child: const Page2(),
       currentChild: context.widget,
     ),
   ),
   Info(
-    'Right to Left',
-    (_) => RightToLeftTranzition(
+    'Pop (bottom to top)',
+    (BuildContext context) => PopTranzition.bottomToTop(
       child: const Page2(),
+      currentChild: context.widget,
+    ),
+  ),
+  Info(
+    'Pop (left to right)',
+    (BuildContext context) => PopTranzition.leftToRight(
+      child: const Page2(),
+      currentChild: context.widget,
+    ),
+  ),
+  Info(
+    'Pop (right to left)',
+    (BuildContext context) => PopTranzition.rightToLeft(
+      child: const Page2(),
+      currentChild: context.widget,
+    ),
+  ),
+  Info(
+    'Pop (top to bottom)',
+    (BuildContext context) => PopTranzition.topToBottom(
+      child: const Page2(),
+      currentChild: context.widget,
     ),
   ),
   Info(
     'Rotate',
     (_) => RotateTranzition(
+      child: const Page2(),
+    ),
+  ),
+  Info(
+    'Rotate (top center)',
+    (_) => RotateTranzition.topCenter(
+      child: const Page2(),
+    ),
+  ),
+  Info(
+    'Rotate (top left)',
+    (_) => RotateTranzition.topLeft(
+      child: const Page2(),
+    ),
+  ),
+  Info(
+    'Rotate (top right)',
+    (_) => RotateTranzition.topRight(
+      child: const Page2(),
+    ),
+  ),
+  Info(
+    'Rotate (bottom center)',
+    (_) => RotateTranzition.bottomCenter(
+      child: const Page2(),
+    ),
+  ),
+  Info(
+    'Rotate (bottom left)',
+    (_) => RotateTranzition.bottomLeft(
+      child: const Page2(),
+    ),
+  ),
+  Info(
+    'Rotate (bottom right)',
+    (_) => RotateTranzition.bottomRight(
+      child: const Page2(),
+    ),
+  ),
+  Info(
+    'Rotate (center left)',
+    (_) => RotateTranzition.centerLeft(
+      child: const Page2(),
+    ),
+  ),
+  Info(
+    'Rotate (center right)',
+    (_) => RotateTranzition.centerRight(
       child: const Page2(),
     ),
   ),
@@ -106,44 +158,50 @@ List<Info> kTransitions = <Info>[
   ),
   Info(
     'Scale (top center)',
-    (_) => ScaleTranzition(
+    (_) => ScaleTranzition.topCenter(
       child: const Page2(),
-      alignment: Alignment.topCenter,
     ),
   ),
   Info(
     'Scale (top left)',
-    (_) => ScaleTranzition(
+    (_) => ScaleTranzition.topLeft(
       child: const Page2(),
-      alignment: Alignment.topLeft,
     ),
   ),
   Info(
     'Scale (top right)',
-    (_) => ScaleTranzition(
+    (_) => ScaleTranzition.topRight(
       child: const Page2(),
-      alignment: Alignment.topRight,
     ),
   ),
   Info(
     'Scale (bottom center)',
-    (_) => ScaleTranzition(
+    (_) => ScaleTranzition.bottomCenter(
       child: const Page2(),
-      alignment: Alignment.bottomCenter,
     ),
   ),
   Info(
     'Scale (bottom left)',
-    (_) => ScaleTranzition(
+    (_) => ScaleTranzition.bottomLeft(
       child: const Page2(),
-      alignment: Alignment.bottomLeft,
     ),
   ),
   Info(
     'Scale (bottom right)',
-    (_) => ScaleTranzition(
+    (_) => ScaleTranzition.bottomRight(
       child: const Page2(),
-      alignment: Alignment.bottomRight,
+    ),
+  ),
+  Info(
+    'Scale (center left)',
+    (_) => ScaleTranzition.centerLeft(
+      child: const Page2(),
+    ),
+  ),
+  Info(
+    'Scale (center right)',
+    (_) => ScaleTranzition.centerRight(
+      child: const Page2(),
     ),
   ),
   Info(
@@ -153,28 +211,38 @@ List<Info> kTransitions = <Info>[
     ),
   ),
   Info(
-    'Top To Bottom Fade',
-    (_) => TopToBottomFadeTranzition(
+    'Size (top)',
+    (_) => SizeTranzition.top(
       child: const Page2(),
     ),
   ),
   Info(
-    'Top To Bottom Join',
-    (BuildContext context) => TopToBottomJoinTranzition(
+    'Size (bottom)',
+    (_) => SizeTranzition.bottom(
       child: const Page2(),
-      currentChild: context.widget,
     ),
   ),
   Info(
-    'Top To Bottom Pop',
-    (BuildContext context) => TopToBottomPopTranzition(
+    'Slide (bottom to top)',
+    (_) => SlideTranzition.bottomToTop(
       child: const Page2(),
-      currentChild: context.widget,
     ),
   ),
   Info(
-    'Top To Bottom',
-    (_) => TopToBottomTranzition(
+    'Slide (left to right)',
+    (_) => SlideTranzition.leftToRight(
+      child: const Page2(),
+    ),
+  ),
+  Info(
+    'Slide (right to left)',
+    (_) => SlideTranzition.rightToLeft(
+      child: const Page2(),
+    ),
+  ),
+  Info(
+    'Slide (top to bottom)',
+    (_) => SlideTranzition.topToBottom(
       child: const Page2(),
     ),
   ),